+++ persons.json
@@ -1,1 +1,1 @@
-[{"json_class":"Student","id":46,"name":"fe","age":12,"parent_permission":true}]
+[{"json_class":"Teacher","id":17,"name":"dfadsf","age":23,"specialization":"fsdaf"},{"json_class":"Student","id":428,"name":"sdfadsf","age":23,"parent_permission":true}]
